# internal/resources/model_request_error_resource.go
/*
 * GENERATED. Do not modify. Your changes might be overwritten!
 */

package resources

type RequestErrorResource struct {
	Detail string `json:"detail"`
	Status string `json:"status"`
	Title  string `json:"title"`
}