# internal/resources/model_update_user.go
/*
 * GENERATED. Do not modify. Your changes might be overwritten!
 */

package resources

type UpdateUser struct {
	Email      *string `json:"email,omitempty"`
	FirstName  *string `json:"first_name,omitempty"`
	Id         *int64  `json:"id,omitempty"`
	Login      *string `json:"login,omitempty"`
	Password   *string `json:"password,omitempty"`
	SecondName *string `json:"second_name,omitempty"`
}